fix backend

+++ controllers/task.routes.js
@@ -4,39 +4,44 @@ const verifyToken = require('../middleware/verify-token')
 
 
 // post router 
-router.post('/',verifyToken, async(req, res)=>{
-    try{
-        req.body.username = req.body._id
-        const createdTask= await Task.create(req.body)
-        res.status(201).json(createdTask)
+router.post('/', verifyToken, async (req, res) => {
+  try {
+    req.body.username = req.user._id
 
+    if (!req.body.category || req.body.category === "") {
+      delete req.body.category
     }
-    catch(err){
-        console.log(err)
-        res.status(500).json(err)
-    }
+
+    const createdTask = await Task.create(req.body)
+    res.status(201).json(createdTask)
+
+  } catch (err) {
+    console.log(err)
+    res.status(500).json(err)
+  }
 })
 
 
 // get router
 
-router.get('/', async (req,res)=>{
-    try{
+router.get('/', verifyToken, async (req, res) => {
+    try {
+        const allTasks = await Task.find({ username: req.user._id })
+            .populate('username')
+            .populate('category')
 
-        const allTasks = await Task.find().populate('username').populate('category')
-        res.json(allTasks)
+        res.status(200).json(allTasks)   
 
-    }
-    catch(err){
+    } catch (err) {
         console.log(err)
         res.status(500).json(err)
     }
 })
 
 // GET /tasks/:id
-router.get('/:id', async(req,res)=>{
+router.get('/:id',verifyToken, async(req,res)=>{
     try{
-        const foundTask = await Task.findById(req.params.id)
+        const foundTask = await Task.findById(req.params.id).populate('username')
         console.log(foundTask)
 
         if(!foundTask){
@@ -55,39 +60,60 @@ router.get('/:id', async(req,res)=>{
 })
 
 // DELETE /tasks/:id
-router.delete('/:id',async(req,res)=>{
-    try{
-        const foudnTask = await Task.findById(req.params.id)
-        console.log(foudnTask)
+router.delete('/:id', verifyToken, async (req, res) => {
+  try {
+    const foundTask = await Task.findById(req.params.id)
 
-        if(!foundTask){
-            return res.status(404).json({message:"No Task with this ID exists"})
-        }
-
-        await Task.findByIdAndDelete(req.params.id)
-
-        res.json({message:"Task successfully deleted"})
-
+    if (!foundTask) {
+      return res.status(404).json({ message: "No Task with this ID exists" })
     }
-    catch(err){
-        console.log(err)
-        res.json(err)
+
+    if (foundTask.username.toString() !== req.user._id) {
+      return res.status(403).json({ message: "Not authorized" })
     }
+
+    await Task.findByIdAndDelete(req.params.id)
+
+    res.json({ message: "Task successfully deleted" })
+  } catch (err) {
+    console.log(err)
+    res.status(500).json(err)
+  }
 })
 
 // PUT /task/:id
 
-router.put('/:id',async(req,res)=>{
-    try{
-        const updatedTask = await Task.findByIdAndUpdate(req.params.id,req.body)
-        res.json(updatedTask)
-    }
-    catch(err){
-        console.log(err)
-        res.json(err)
-    }
-})
+router.put('/:id', verifyToken, async (req, res) => {
+  try {
+    const foundTask = await Task.findById(req.params.id)
 
+    if (!foundTask) {
+      return res.status(404).json({ message: "No Task with this ID exists" })
+    }
+
+    if (foundTask.username.toString() !== req.user._id) {
+      return res.status(403).json({ message: "Not authorized" })
+    }
+
+
+    const updateData = { ...req.body }
+
+    if (!updateData.category) delete updateData.category
+    if (updateData.category === "") delete updateData.category
+
+    const updatedTask = await Task.findByIdAndUpdate(
+      req.params.id,
+      updateData,
+      { new: true, runValidators: true }
+    )
+
+    res.json(updatedTask)
+
+  } catch (err) {
+    console.log(err)
+    res.status(500).json(err)
+  }
+})
 
 
 module.exports = router
